tests/submission/Microsoft/semantic/article.html

<!doctype html>
<html>
 <head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type" />
  <title>Verify ARTICLE is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="http://www.w3c-test.org/resources/testharnessreport.js"></script>
</head>
<body>
 <div id="log"></div>
 <article id="article">
 <header>
  <h1>H1 in an article and header</h1>
 </header>
 </article>

<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("article").toString() )
 })
 </script>
</body>
</html>
